refactor: 调整Base类层级
