+++ src/main/java/moe/takochan/takotech/common/block/IBaseAETileBlock.java
@@ -1,0 +1,5 @@
+package moe.takochan.takotech.common.block;
+
+public interface IBaseAETileBlock extends IBaseBlock {
+
+}
