--- src/main/java/moe/takochan/takotech/common/block/ae/IBaseAETileBlock.java
@@ -1,7 +1,0 @@
-package moe.takochan.takotech.common.block.ae;
-
-import moe.takochan.takotech.common.block.IBaseBlock;
-
-public interface IBaseAETileBlock extends IBaseBlock {
-
-}
