+++ src/main/java/moe/takochan/takotech/common/item/ae/ItemOreStorageCell.java
@@ -33,6 +33,7 @@ import cpw.mods.fml.common.registry.GameRegistry;
 import cpw.mods.fml.relauncher.Side;
 import cpw.mods.fml.relauncher.SideOnly;
 import moe.takochan.takotech.client.tabs.TakoTechTabs;
+import moe.takochan.takotech.common.item.BaseAECellItem;
 import moe.takochan.takotech.common.storage.ITakoCellInventory;
 import moe.takochan.takotech.common.storage.ITakoCellInventoryHandler;
 import moe.takochan.takotech.common.storage.inventory.OreStorageCellInventory;
